templates/connectors/gmail/files/.slice/plugins/google-workspace/tools/google_workspace_mcp/pyproject.toml
[project]
name = "life-google-workspace-mcp"
version = "0.1.0"
description = "Read-only Google Calendar and Gmail MCP server for Slice."
requires-python = ">=3.10"
dependencies = [
  "google-api-python-client>=2.0.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.0",
  "mcp[cli]>=1.0.0",
]

[project.scripts]
google-workspace-auth = "google_workspace_mcp.auth:main"
google-workspace-mcp = "google_workspace_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["google_workspace_mcp"]
